[ADDS] comments and small changes to clean up GUI code.
