+++ Ravel-Creators/Assets/Ravel-Creators/Scripts/Editor/Environment/EnvironmentSOEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5fb95176c1ad62c419700bac7e5cae9a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
