+++ Ravel-Creators/Assets/Ravel-Creators/Scripts/Editor/RavelWindow/CreatorWindow.cs
@@ -4,12 +4,19 @@ using UnityEditor;
 using UnityEngine;
 using MathBuddy.Strings;
 
+/// <summary>
+/// Parent of the creator window. These gui calls happen for each of the tabs in the window.
+/// The tabs are also called through this class.
+/// </summary>
 public class CreatorWindow : EditorWindow
 {
+	/// <summary>
+	/// Currently open tab.
+	/// </summary>
 	public CreatorWindowState Tab {
 		get {
 			if (!_states.ContainsKey(_tab)) {
-				_states.Add(_tab, GetState(_tab));
+				_states.Add(_tab, CreateState(_tab));
 			}
 			return _states[_tab];
 		}
@@ -19,11 +26,6 @@ public class CreatorWindow : EditorWindow
 	
 	private State _tab;
 	private Dictionary<State, CreatorWindowState> _states = new Dictionary<State, CreatorWindowState>();
-
-	//[MenuItem("Ravel/Creator", false, 0)]
-	public static void OpenWindow() {
-		GetWindow();
-	}
 	
 	[MenuItem("Ravel/Creator/Account", false, 1)]
 	public static void OpenAccount() {
@@ -35,11 +37,14 @@ public class CreatorWindow : EditorWindow
 		GetWindow(State.Environments);
 	}
 
-	public static CreatorWindow GetWindow(State s = State.None, bool show = true) {
+	/// <summary>
+	/// Either creates or retrieves reference to the creator window.
+	/// </summary>
+	/// <param name="tab">tab on which to open the creator window.</param>
+	/// <param name="show">Should the window be opened, or just retrieved.</param>
+	public static CreatorWindow GetWindow(State tab, bool show = true) {
 		CreatorWindow wnd = GetWindow<CreatorWindow>();
-		if (s != State.None) {
-			wnd.SetState(s);
-		}
+		wnd.SwitchTab(tab);
 		
 		if (show) {
 			wnd.Show();	
@@ -47,51 +52,56 @@ public class CreatorWindow : EditorWindow
 		return wnd;
 	}
 
-	public void SetState(State newState) {
-		_tab = newState;
+	/// <summary>
+	/// Switch tab of window.
+	/// </summary>
+	public void SwitchTab(State newTab) {
+		_tab = newTab;
 	}
 
-	public CreatorWindowState GetState(State s) {
-		switch (s) {
+	/// <summary>
+	/// Creates a new instance of a tab/state
+	/// </summary>
+	/// <param name="tab">state to create</param>
+	public CreatorWindowState CreateState(State tab) {
+		switch (tab) {
 			case State.Account:
 				return new AccountState(this);
 			case State.Environments:
 				return new EnvironmentState(this);
 			default:
-				throw new Exception($"Missing creator window state ({s})");
+				throw new Exception($"Missing creator window state ({tab})");
 		}
 	}
 
 	private void OnGUI() {
 		if (!RavelEditor.LoggedIn) {
-			_tab = State.Account;
+			SwitchTab(State.Account);
+			//cannot switch tabs when not logged in
 			GUI.enabled = false;
 		}
 		
 		_tab = (State)GUILayout.Toolbar((int)_tab, GetStateNames());
+		//if it was disabled, re-enable gui
 		GUI.enabled = true;
 		
 		if (RavelEditor.Branding.banner) {
 			RavelEditor.DrawTextureScaledCropGUI(new Rect(0, GUILayoutUtility.GetLastRect().yMax, position.width, RavelBranding.BANNER_HEIGHT), 
 				RavelEditor.Branding.banner, RavelEditor.Branding.bannerPOI);
 		}
-
 		
-		if (_tab != State.None) {
-			_scroll = EditorGUILayout.BeginScrollView(_scroll, GUILayout.Width(position.width));
-			Tab.OnGUI(position);
-		}
-		
-		if (RavelEditor.Branding.daan) {
-			RavelEditor.DrawTextureScaledCropGUI(new Rect(position.width / 2f - 50,  GUILayoutUtility.GetLastRect().yMax, 100f, 100f), 
-				RavelEditor.Branding.daan, Vector2.one * 0.5f);
-		}
-
-		if (_tab != State.None) {
-			EditorGUILayout.EndScrollView();
-		}
+		_scroll = EditorGUILayout.BeginScrollView(_scroll, GUILayout.Width(position.width));
+		Tab.OnGUI(position);
+			
+		RavelEditor.DrawTextureScaledCropGUI(new Rect(position.width / 2f - 50,  GUILayoutUtility.GetLastRect().yMax, 100f, 100f), 
+			RavelEditor.Branding.daan, Vector2.one * 0.5f);
+			
+		EditorGUILayout.EndScrollView();
 	}
-
+	
+	/// <summary>
+	/// Re-formats the state enum below into more readable names, so they can be shown as the names of the tabs.
+	/// </summary>
 	private string[] GetStateNames() {
 		string[] names = Enum.GetNames(typeof(State));
 		for (int i = 0; i < names.Length; i++) {
@@ -101,10 +111,12 @@ public class CreatorWindow : EditorWindow
 
 		return names;
 	}
-
+	
+	/// <summary>
+	/// Tabs for the creator window. New tabs also have to be added to the CreateState method.
+	/// </summary>
 	public enum State
 	{
-		None = 0,
 		Account,
 		Environments,
 	}
