+++ Ravel-Creators/Assets/Ravel-Creators/Scripts/Editor/RavelWindow/States/CreatorWindowState.cs
@@ -1,7 +1,13 @@
 using UnityEngine;
 
+/// <summary>
+/// Base class for sub-windows of the Creator window. 
+/// </summary>
 public abstract class CreatorWindowState
 {
+	/// <summary>
+	/// State identifier of this window.
+	/// </summary>
 	public abstract CreatorWindow.State State { get; } 
 
 	private CreatorWindow _wnd;
@@ -12,11 +18,17 @@ public abstract class CreatorWindowState
 	
 	public abstract void OnGUI(Rect position);
 
+	/// <summary>
+	/// Opens creator window to this sub-page.
+	/// </summary>
 	public void Show() {
-		_wnd.SetState(State);
+		_wnd.SwitchTab(State);
 		_wnd.Show();
 	}
 
+	/// <summary>
+	/// Closes the creator window
+	/// </summary>
 	public void Close() {
 		_wnd.Close();
 	}
