+++ Ravel-Creators/Assets/Ravel-Creators/Scripts/Editor/WebRequests/CreatorRequest.cs
@@ -5,6 +5,9 @@ using Newtonsoft.Json;
 using UnityEngine;
 using File = UnityEngine.Windows.File;
 
+/// <summary>
+/// Creator webrequests shown on the API documentation under tab Environments.
+/// </summary>
 public class CreatorRequest : TokenWebRequest
 {
     public CreatorRequest(Method method, string postfix, string version = "v1/") : base(method, "api/", version) {
@@ -21,6 +24,11 @@ public class CreatorRequest : TokenWebRequest
         _url += "environments/" + postfix;
     }
 
+    /// <summary>
+    /// Creates a new environment on server.
+    /// </summary>
+    /// <param name="userUuid">UUID of the user that creates the environment.</param>
+    /// <param name="env">environment which will be created (contains name, descriptions and isPublic)</param>
     public static CreatorRequest CreateEnvironment(string userUuid, Environment env) {
         string json = JsonConvert.SerializeObject(env);
         json = EnvironmentExtensions.RenameStringToBackend(json);
@@ -28,6 +36,10 @@ public class CreatorRequest : TokenWebRequest
         return new CreatorRequest(Method.PostJSON, $"{userUuid}", json, "v1/");
     }
     
+    /// <summary>
+    /// Deletes pushed environment from the server (only works for unpublished environments).
+    /// </summary>
+    /// <param name="env">environment that has to be deleted.</param>
     public static CreatorRequest DeleteEnvironment(Environment env) {
         if (env.published) {
             throw new Exception("Cannot delete published environments");
@@ -35,10 +47,17 @@ public class CreatorRequest : TokenWebRequest
         return new CreatorRequest(Method.Delete, $"{env.environmentUuid}");
     }
     
+    /// <summary>
+    /// Publishes an environment (without review)
+    /// </summary>
+    /// <param name="envUuid">UUID of environment that will be pushed.</param>
     public static CreatorRequest PublishEnvironment(string envUuid) {
-        return new CreatorRequest(Method.Put, $"submissions/{envUuid}", "v1/");
+        return new CreatorRequest(Method.Put, $"publish/{envUuid}", "v2/");
     }
     
+    /// <summary>
+    /// Uploads a image (located at path) as preview for environment with given UUID.
+    /// </summary>
     public static CreatorRequest UploadPreview(string envUuid, string imagePath) {
         WWWForm form = new WWWForm();
         string fName = Path.GetFileName(imagePath);
@@ -47,6 +66,9 @@ public class CreatorRequest : TokenWebRequest
         return new CreatorRequest($"uploads/preview-images?environmentUuid={envUuid}", form);
     }
     
+    /// <summary>
+    /// Uploads an assetbundle (located at path) as bundle for environment with given UUID.
+    /// </summary>
     public static CreatorRequest UploadBundle(string envUuid, string bundlePath) {
         WWWForm form = new WWWForm();
         string fName = Path.GetFileName(bundlePath);
@@ -55,12 +77,20 @@ public class CreatorRequest : TokenWebRequest
         return new CreatorRequest($"uploads/asset-bundles?environmentUuid={envUuid}", form);
     }
     
+    /// <summary>
+    /// Get all environments for a given creator.
+    /// </summary>
+    /// <param name="userUuid">user UUID of user.</param>
+    /// <param name="isPublished">Should the published or unpublished environments be pushed.</param>
     public static CreatorRequest GetCreatorEnvironments(string userUuid, bool isPublished) {
         CreatorRequest req = new CreatorRequest(Method.Get, $"{userUuid}");
         req.AddParameter("isPublished", isPublished.ToString());
         return req;
     }
     
+    /// <summary>
+    /// Get one specific creator environment based on it's UUID
+    /// </summary>
     public static CreatorRequest GetCreatorEnvironment(string envUuid) {
         return new CreatorRequest(Method.Get, $"single/{envUuid}");
     }
